#77 feat: 백그라운드 걸음 수 캐싱 구현
