--- Walkie-iOS/Walkie-iOS/Sources/Domain/UseCase/Egg/Implementation/DefaultHatchEggUseCase.swift
@@ -1,21 +1,0 @@
-//
-//  DefaultHatchEggUseCase.swift
-//  Walkie-iOS
-//
-//  Created by 황채웅 on 5/1/25.
-//
-
-import Foundation
-import SwiftUICore
-
-final class DefaultHatchEggUseCase: BaseEggUseCase, HatchEggUseCase {
-    
-    func hatchEgg() {
-        switch UserManager.shared.getCharacterType {
-        case .dino:
-            let character = UserManager.shared.getDinoType
-        case .jellyfish:
-            let character = UserManager.shared.getJellyfishType
-        }
-    }
-}
