--- Walkie-iOS/Walkie-iOS/Sources/Domain/UseCase/Egg/Protocol/HatchEggUseCase.swift
@@ -1,12 +1,0 @@
-//
-//  HatchEggUseCase.swift
-//  Walkie-iOS
-//
-//  Created by 황채웅 on 5/1/25.
-//
-
-import Foundation
-
-protocol HatchEggUseCase {
-    func hatchEgg()
-}
