+++ Walkie-iOS/Walkie-iOS/Sources/Domain/UseCase/Step/Protocol/UpdateStepUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  UpdateStepUseCase.swift
+//  Walkie-iOS
+//
+//  Created by 황채웅 on 5/3/25.
+//
+
+import Foundation
+
+protocol UpdateStepCacheUseCase {
+    func execute()
+}
